rad/kod/ecu-template/requirements.txt
can-isotp==2.0.4
msgpack==1.0.8
packaging==24.0
python-can==4.3.1
typing_extensions==4.11.0
wrapt==1.16.0
